+++ src/web/renderer/canvas-renderer/canvas-renderer.js
@@ -149,7 +149,7 @@ export function renderer(sceneFn = sceneFactory) {
     return config.Promise.resolve(doRender);
   };
 
-  canvasRenderer.itemsAt = options => scene.getItemsFrom(options);
+  canvasRenderer.itemsAt = options => (scene ? scene.getItemsFrom(options) : []);
 
   canvasRenderer.clear = () => {
     if (!el) {